+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/ApplicationApprovalService_New.java
@@ -9,6 +9,6 @@ public interface ApplicationApprovalService_New {
 	
 	public void insert(Application application);
 	
-	public void delete(String companyID, String appID, Long version, ApplicationType_Old appType);
+	public void delete(String companyID, String appID, int version, ApplicationType_Old appType);
 	
 }
